% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actnorm.R
\name{actnorm}
\alias{actnorm}
\title{actnorm : The probability of an event when all subjects are assigned to the active treatment}
\usage{
actnorm(coeff, dpar)
}
\arguments{
\item{coeff}{The coefficients obtained from fitting the log linear model and must be a vector of length four}

\item{dpar}{The parameters of the biomarker which has a normal distribution and must be of a vector legnth two}
}
\value{
A value of the estimated probability and must be a number between 0 and 1
}
\description{
This function is used to estimate the probability of an event assuming all the subjects are
             are assigned the treatment group. The predictive biomarker is assumed
             to have a normal distribution with mean nu and variance sigma squared.
}
\examples{
  # Let coff is a vector of length four with each elemnt represent the value of the coefficients of the model
  # Let dpar is a vector of legnth two such that the  mean= 2.3 and variance = 0.2 of the predictive biomarker are given
  eventactive <- actnorm(coff, dpar=c(2.3,0.2))
}
\author{
Henok Woldu
}
