% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl2mp.R
\name{cl2mp}
\alias{cl2mp}
\title{cl2mp : Change Clinician Input to MOdel parameters}
\usage{
cl2mp(k, z)
}
\arguments{
\item{k}{A vector of length four which specify the clinician input values.}

\item{z}{A vector of length two for the 25th and 75th percentile value of the biomarker values.}
}
\value{
A vector of lenght four containing the coefficients of the model
}
\description{
The function is used to change the clinician input values to model parameters.
This values are to be the best intelectural guess from the expertise in that
area of study or they can be taken from previous studies.
}
\examples{
  # If we know the 25th and 75th percentile values of the biomarker are 3.2 and 5.8
  # And K1 = 0.6 , K2 = 2.5 , K3= 1.5 and K4= 0.6, then use the function like
  betavalues <- cl2mp(c(k1,k2,k3,k4), c(3.2,5.8))
  betavalues
}
\author{
Henok Woldu
}
