% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datgen.R
\name{datgen}
\alias{datgen}
\title{datgen : Function to generate data}
\usage{
datgen(m, n, coeff, dpar)
}
\arguments{
\item{m}{The number of data sets to generate}

\item{n}{The sample size of each data set}

\item{dpar}{The parameter of the biomarker distribution}

\item{coff}{The initial coefficients of the model to generate data}
}
\value{
A list with m data set each with n sample size
}
\description{
This is used to generate a data from a logit model. The out come is a binary which represents
the whether the event is observed or not. The treatment arm, biomarker value are the main
exposure variables.
}
\examples{
  # Lets say we want to generat m =100 data set each with n= 500 sample size
  # Further lets assume the biomarker is normally distributed with mean 0 and var = 1
  # If the var of the error is given to be 0.1, then
  mydat <- datgen(100,500,c(0,1,0.1))
}
\author{
Henok Woldu
}
