% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta4r.R
\name{delta4r}
\alias{delta4r}
\title{delta4r : Estimated reproducibility metric}
\usage{
delta4r(mydatt, dpar)
}
\arguments{
\item{dpar}{The parameters of the normally distributed biomarker.}

\item{mydat}{A data frame with the outcome variable and exposure variables}
}
\value{
A value in the range  of 0-1.
}
\description{
This function is used the reproducibility metric. It is the difference in
             theta
}
\details{
This metric measure the difference in the decrease of  expected proportion of events when the
         gold standard bimarker is measured and when the modified asssay is measured.
}
\examples{
  # Let the data is stored in an object named mydat
  # Let dpar the parmaters of the normally distributed biomarker
  truetheta <- thetags(mydat, dpar)
}
\author{
Henok Woldu
}
