% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetags.R
\name{thetags}
\alias{thetags}
\title{thetags : Estimated theta when the gold standard biomarker is observed}
\usage{
thetags(mydatt, dpar)
}
\arguments{
\item{dpar}{The parameters of the normally distributed biomarker.}

\item{mydat}{A data frame with the outcome variable and exposure variables}
}
\value{
A value in the range  of 0-1.
}
\description{
This function is used to estimate theta assuming the gold standard biomarker is used.
             Further the distribution of the biomarker is assumed to be normaly distributed with
             a given mean and variance.
}
\details{
This metric measure the decrease in the expected proportion of events that were avoided
         as a result of using the biomarker guided treatment. This is a direct biomarker utility
         metric as can quantify the net benefit that we can gain using the biomarker to guide treatmetn.
}
\examples{
  # Let the data is stored in an object named mydat
  # Let dpar the parmaters of the normally distributed biomarker
  truetheta <- thetags(mydat, dpar)
}
\author{
Henok Woldu
}
