% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetama.R
\name{thetama}
\alias{thetama}
\title{thetama : Estimated theta when the modified assay is observed.}
\usage{
thetama(mydat)
}
\arguments{
\item{mydat}{A data frame with the outcome variable and exposure variables}
}
\value{
A value in the range  of 0-1.
}
\description{
This function is used to estimate theta assuming the gold standard biomarker is used.
             Further the distribution of the biomarker is assumed to be normaly distributed with
             a given mean and variance.
}
\details{
This metric measure the decrease in the expected proportion of events that were avoided
         as a result of using the biomarker guided treatment. However the biomarker observed is the
         modified asssay rather than the gold standard assay. To reduce the computational time, a raoblackwelization
         method conditioning over the sufficient statistics was used.
}
\examples{
  # Let the data is stored in an object named mydat
  # Let dpar the parmaters of the normally distributed biomarker
  modtheta <- thetama(mydat)
}
\author{
Henok Woldu
}
